Stop using CommandScheduler hacks
Instead of creating new commands to switch speed controller outputs, have the PrototypeSubsystem handle all routing of outputs depending on the state of the system (controlled through a Shuffleboard chooser). This leaves PrototypeControlCommands to be parameterized only by the PrototypeSubsystem object and the output speed value source, eliminating the need to create new commands to switch speed controllers. A singular PrototypeControlCommand can be bound to a button, and there is no need to manually schedule commands (CommandScheduler hacks).